+++ lib/eda/attachment.ex
@@ -70,7 +70,7 @@ defmodule EDA.Attachment do
           waveform: String.t() | nil,
           flags: integer() | nil,
           clip_participants: [EDA.User.t()] | nil,
-          clip_created_at: String.t() | nil,
+          clip_created_at: DateTime.t() | nil,
           application: map() | nil
         }
 
@@ -96,7 +96,7 @@ defmodule EDA.Attachment do
       waveform: raw["waveform"],
       flags: raw["flags"],
       clip_participants: parse_users(raw["clip_participants"]),
-      clip_created_at: raw["clip_created_at"],
+      clip_created_at: EDA.Timestamp.parse(raw["clip_created_at"]),
       application: raw["application"]
     }
   end