+++ test/eda/kept_fields_test.exs
@@ -37,7 +37,7 @@ defmodule EDA.KeptFieldsTest do
       })
 
     assert [%EDA.User{username: "streamer"}] = attachment.clip_participants
-    assert attachment.clip_created_at == "2026-09-23T10:00:00.000000+00:00"
+    assert attachment.clip_created_at == ~U[2026-09-23 10:00:00.000000Z]
     assert %{"name" => "Some Game"} = attachment.application
   end