+++ test/eda/poll_test.exs
@@ -29,7 +29,7 @@ defmodule EDA.PollTest do
       poll = Poll.from_raw(raw)
       assert poll.question == "Best language?"
       assert length(poll.answers) == 2
-      assert poll.expiry == "2025-06-01T00:00:00+00:00"
+      assert poll.expiry == ~U[2025-06-01 00:00:00Z]
       assert poll.allow_multiselect == true
       assert poll.layout_type == 1